\name{dbMetaLoader-package}
\alias{dbMetaLoader-package}
\alias{dbMetaLoader}
\docType{package}

\title{
Scripts to automate and extend the loading of data into PostGIS.
}

\description{
There are many tools to load GIS data into a PostGIS database, though none quite fitted what I was after. My approach here is focused on automation and the collection of metadata to provide context and licensing information as required.
This packages is mainly intended for my own use and as a learning tool for myself, but is open for anyone to use/modify if they want.
}

\details{
\tabular{ll}{
Package: \tab dbMetaLoader\cr
Type: \tab Package\cr
Version: \tab 0.1-0\cr
Date: \tab 2016-08-08\cr
License: \tab GPL3\cr
}

}
\author{
Joseph McGrath <joseph.mcgrath@outlook.com>
}

\references{
Listed with individual function.
}

\keyword{
package
}

