\name{spatialTables}
\alias{spatialTables}

\title{
Produces a RegEx string that can be used to find spatial data.
}

\description{
Produces a string that looks for file extensions associated with spatial data (e.g. ESRI shapefiles, MapInfo TAB files).
}

\usage{
spatialTables()
}

\arguments{
  \item{No arguments}{This function has no arguments.}
}

\value{Returns a RegEx string.}

\author{Joseph McGrath}
